fix cross-building of the Linux kernel module
In case one cross-builds the tpm_emulator the Linux kernel module was compiled
using the host compiler, which would fail for obvious reasons. It was further
impossible to tell the build to use a specific kernel source directory for both
native and cross builds.

The logic is now changed as follows when building on Linux:
-if the compiler is gcc then the compiler used for the application will also be
 used when building the kernel module
-for any other compiler the user is free to pass CC= and LD= to the make call
-the user can (and should) pass a ARCH= variable to the make call to specify
 the target architecture of the kernel to be used

The call for a successful cross-build from x86-32 to ARM looked like this:

  mkdir build
  cd build
  cmake -D CMAKE_TOOLCHAIN_FILE=/sysroot/toolchain.cmake \
        -D CMAKE_INSTALL_PREFIX=/usr ../tpm_emulator
  make KERNEL_BUILD=/sysroot/src/linux-3.2.64 ARCH=arm

Signed-off-by: Rolf Eike Beer <eb@emlix.com>