Use microseconds for timeouts and durations
The timeout and duration fields should be in microseconds according to
the spec.

TPM_CAP_PROP_TIS_TIMEOUT:
A 4 element array of UINT32 values each denoting the timeout value in
microseconds for the following in this order:

TPM_CAP_PROP_DURATION:
A 3 element array of UINT32 values each denoting the duration value in
microseconds of the duration of the three classes of commands:

Linux will scale the timeouts up by 1000, but not the durations.  Change
the units for both sets as appropriate.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>